# perf: double directory size calculation speed