--- a/yazi-fs/src/lib.rs
+++ b/yazi-fs/src/lib.rs
@@ -2,7 +2,7 @@
 
 yazi_macro::mod_pub!(cha mounts);
 
-yazi_macro::mod_flat!(cwd file files filter fns op path sorter sorting stage step xdg);
+yazi_macro::mod_flat!(calculator cwd file files filter fns op path sorter sorting stage step xdg);
 
 pub fn init() {
 	CWD.init(<_>::default());