This PR greatly improves directory computation performance by dividing the task into chunks and reusing existing threads. Directory sorting and file trashing will be observably twice as fast.

By leveraging chunking, it now can produce real-time directory statistics reports. This means that for a large directory, you will be able to see the process of size changes in real-time, which is beneficial for the upcoming `fs.dir_size()` API.

The new implementation is still async cancelable, which is important for using `fs.dir_size()` in spotters, as it will allow canceling a still-incomplete directory computation when quickly switching between files, thereby avoiding wasted resources.

## Benchmark

```rust
let p = std::path::Path::new("/tmp/yazi");

let now = std::time::Instant::now();
yazi_fs::calculate_size(p).await;
println!("Old: elapsed {:?}", now.elapsed());

let now = std::time::Instant::now();
yazi_fs::SizeCalculator::total(p).await?;
println!("New: elapsed {:?}", now.elapsed());
```

Run the benchmark script against a 10.16GB folder 5 times in release mode:

```sh
Old: elapsed 759.877791ms
New: elapsed 364.802583ms

Old: elapsed 734.5155ms
New: elapsed 360.103375ms

Old: elapsed 734.124167ms
New: elapsed 365.662458ms

Old: elapsed 733.27325ms
New: elapsed 365.694042ms

Old: elapsed 743.394875ms
New: elapsed 364.455083ms
```